+++ lib/worklist/types.ts
@@ -117,3 +117,27 @@ export interface SuggestedMatch {
   counterparty_name: string | null
   candidate_total: number | null
 }
+
+/**
+ * Journal-entry source types that require underlag (BFL 5 kap 7 §). Source
+ * types representing system-generated entries (VAT settlement, year-end,
+ * currency revaluation, ...) are exempt by omission.
+ *
+ * Single source of truth for EVERY TS surface (worklist counts, journal-list
+ * chip/waiver UI, no-doc-required batch route, push notifications); the SQL
+ * mirror lives in the verifikat_without_documents RPC, pinned by
+ * tests/pg/document-surfaces-unification.pg.test.ts. Lives here (not in
+ * categories.ts) because this module is dependency-free and safe to import
+ * from client components.
+ */
+export const NEEDS_DOC_SOURCE_TYPES = [
+  'manual',
+  'bank_transaction',
+  'supplier_invoice_registered',
+  'supplier_invoice_paid',
+  'supplier_invoice_cash_payment',
+  'import',
+  // Webshop order bookings rest on the generated orderunderlag (#1881); an
+  // entry whose underlag failed to attach must surface here.
+  'webshop_order',
+] as const
